:bug: fix: export authOptions in route
