+++ src/app/api/auth/[...nextauth]/options.ts
@@ -1,0 +1,26 @@
+import { type AuthOptions } from "next-auth";
+import GoogleProvider from "next-auth/providers/google";
+import { PrismaAdapter } from "@next-auth/prisma-adapter";
+import prisma from "@/lib/prismadb";
+
+const authOptions: AuthOptions = {
+  adapter: PrismaAdapter(prisma),
+  debug: process.env.NODE_ENV === "development" ? true : false,
+  session: {
+    strategy: "jwt",
+  },
+  secret: process.env.NEXTAUTH_SECRET,
+  pages: {
+    signIn: "/auth/login",
+    signOut: "/auth/signout",
+  },
+  providers: [
+    GoogleProvider({
+      clientId: process.env.GOOGLE_CLIENT_ID!,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
+    }),
+  ],
+  callbacks: {},
+};
+
+export default authOptions;
